Some more cases for propagating change to workspace
Also, the workspace item was removed because we are using the
LocalItemInfo for info about workspace items.  We also no longer need
WorkspaceListener because the concept of listening to workspace changes
no longer exists.  Instead, changes are managed within the Workspace
class and it is just responsible for giving a prioritized list of
suggested files/dirs to try to process (which does not have to be
complete in the case where the list may get really long).